app/requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
pandas==2.2.2
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.3
jinja2==3.1.4
python-multipart==0.0.5
